+++ Realm/RLMArrayLinkView.mm
@@ -173,6 +173,24 @@ inline id RLMCreateAccessorForArrayIndex(RLMArrayLinkView *array, NSUInteger ind
                                    reason:@"Not yet implemented" userInfo:nil];
 }
 
+- (NSUInteger)indexOfObject:(RLMObject *)object {
+    RLMLinkViewArrayValidateAttached(self);
+
+    if (object->_row.get_table() != &_backingLinkView->get_parent()) {
+        @throw [NSException exceptionWithName:@"RLMException"
+                                       reason:@"Object type does not match RLMArray"
+                                     userInfo:nil];
+    }
+
+    size_t object_ndx = object->_row.get_index();
+    size_t result = _backingLinkView->find(object_ndx);
+    if (result == tightdb::not_found) {
+        return NSNotFound;
+    }
+
+    return result;
+}
+
 @end